UPDATE: renamed patters to variations
