+++ src/editor.scss
@@ -1,6 +1,4 @@
 /**
  * The following styles get applied inside the editor only.
- *
- * Replace them with your own styles or remove the file completely.
  */
