+++ src/index.js
@@ -20,7 +20,7 @@ import './style.scss';
 import Edit from './edit';
 import save from './save';
 import metadata from './block.json';
-import './patterns';
+import './variations';
 
 /**
  * Every block starts by registering a new block type definition.
