+++ src/variations/class-diagram.js
