+++ src/variations/sequence-diagram.js
