+++ src/variations/state-diagram.js
